+++ ui/src/dashboards/actions/cellEditorOverlay.ts
@@ -1,3 +1,25 @@
+// Libraries
+import uuid from 'uuid'
+
+// Utils
+import {
+  fill,
+  timeShift,
+  chooseTag,
+  groupByTag,
+  removeFuncs,
+  groupByTime,
+  toggleField,
+  chooseNamespace,
+  chooseMeasurement,
+  addInitialField,
+  applyFuncsToField,
+  toggleTagAcceptance,
+} from 'src/utils/queryTransitions'
+import {getDeep} from 'src/utils/wrappers'
+import defaultQueryConfig from 'src/utils/defaultQueryConfig'
+
+// Types
 import {ColorNumber, ColorString} from 'src/types/colors'
 import {
   DecimalPlaces,
@@ -7,12 +29,28 @@ import {
   CellType,
   ThresholdType,
   TableOptions,
+  CellQuery,
+  NewDefaultCell,
 } from 'src/types/dashboards'
-import {NewDefaultCell} from 'src/types/dashboards'
+import {
+  Status,
+  Field,
+  GroupBy,
+  Tag,
+  TimeShift,
+  ApplyFuncsToFieldArgs,
+} from 'src/types'
+import {CEOInitialState} from 'src/dashboards/reducers/cellEditorOverlay'
+
+interface State {
+  cellEditorOverlay: CEOInitialState
+}
+
+type GetState = () => State
 
 export enum ActionType {
-  ShowCellEditorOverlay = 'SHOW_CELL_EDITOR_OVERLAY',
-  HideCellEditorOverlay = 'HIDE_CELL_EDITOR_OVERLAY',
+  LoadCellForCEO = 'LOAD_CELL_FOR_CEO',
+  ClearCellFromCEO = 'CLEAR_CELL_FROM_CEO',
   ChangeCellType = 'CHANGE_CELL_TYPE',
   RenameCell = 'RENAME_CELL',
   UpdateThresholdsListColors = 'UPDATE_THRESHOLDS_LIST_COLORS',
@@ -24,11 +62,13 @@ export enum ActionType {
   ChangeTimeFormat = 'CHANGE_TIME_FORMAT',
   ChangeDecimalPlaces = 'CHANGE_DECIMAL_PLACES',
   UpdateFieldOptions = 'UPDATE_FIELD_OPTIONS',
+  UpdateQueryDrafts = 'UPDATE_QUERY_DRAFTS',
+  UpdateQueryDraft = 'UPDATE_QUERY_DRAFT',
 }
 
 export type Action =
-  | ShowCellEditorOverlayAction
-  | HideCellEditorOverlayAction
+  | LoadCellForCEOAction
+  | ClearCellFromCEOAction
   | ChangeCellTypeAction
   | RenameCellAction
   | UpdateThresholdsListColorsAction
@@ -40,16 +80,17 @@ export type Action =
   | ChangeTimeFormatAction
   | ChangeDecimalPlacesAction
   | UpdateFieldOptionsAction
+  | UpdateQueryDraftsAction
 
-export interface ShowCellEditorOverlayAction {
-  type: ActionType.ShowCellEditorOverlay
+export interface LoadCellForCEOAction {
+  type: ActionType.LoadCellForCEO
   payload: {
     cell: Cell | NewDefaultCell
   }
 }
 
-export interface HideCellEditorOverlayAction {
-  type: ActionType.HideCellEditorOverlay
+export interface ClearCellFromCEOAction {
+  type: ActionType.ClearCellFromCEO
 }
 
 export interface ChangeCellTypeAction {
@@ -129,17 +170,31 @@ export interface UpdateFieldOptionsAction {
   }
 }
 
-export const showCellEditorOverlay = (
+export interface UpdateQueryDraftsAction {
+  type: ActionType.UpdateQueryDrafts
+  payload: {
+    queryDrafts: CellQuery[]
+  }
+}
+
+export interface UpdateQueryDraftAction {
+  type: ActionType.UpdateQueryDraft
+  payload: {
+    queryDraft: CellQuery
+  }
+}
+
+export const loadCellForCEO = (
   cell: Cell | NewDefaultCell
-): ShowCellEditorOverlayAction => ({
-  type: ActionType.ShowCellEditorOverlay,
+): LoadCellForCEOAction => ({
+  type: ActionType.LoadCellForCEO,
   payload: {
     cell,
   },
 })
 
-export const hideCellEditorOverlay = (): HideCellEditorOverlayAction => ({
-  type: ActionType.HideCellEditorOverlay,
+export const clearCellFromCEO = (): ClearCellFromCEOAction => ({
+  type: ActionType.ClearCellFromCEO,
 })
 
 export const changeCellType = (cellType: CellType): ChangeCellTypeAction => ({
@@ -234,3 +289,332 @@ export const updateFieldOptions = (
     fieldOptions,
   },
 })
+
+// ---------- Query Builder ----------
+
+export const updateQueryDrafts = (
+  queryDrafts: CellQuery[]
+): UpdateQueryDraftsAction => ({
+  type: ActionType.UpdateQueryDrafts,
+  payload: {
+    queryDrafts,
+  },
+})
+
+export const updateQueryDraft = (
+  queryDraft: CellQuery
+): UpdateQueryDraftAction => ({
+  type: ActionType.UpdateQueryDraft,
+  payload: {
+    queryDraft,
+  },
+})
+
+export const addQueryAsync = (queryID: string = uuid.v4()) => async (
+  dispatch,
+  getState: GetState
+) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const newQueryDraft: CellQuery = {
+    query: '',
+    queryConfig: defaultQueryConfig({id: queryID}),
+    source: '',
+    id: queryID,
+  }
+  const updatedQueryDrafts = [...queryDrafts, newQueryDraft]
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const deleteQueryAsync = (queryID: string) => async (
+  dispatch,
+  getState: GetState
+) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.filter(query => query.id !== queryID)
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const toggleFieldAsync = (queryID: string, fieldFunc: Field) => async (
+  dispatch,
+  getState: GetState
+) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = {
+        ...toggleField(query.queryConfig, fieldFunc),
+        rawText: null,
+      }
+
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const groupByTimeAsync = (queryID: string, time: string) => (
+  dispatch,
+  getState: GetState
+) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = groupByTime(query.queryConfig, time)
+
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const fillAsync = (queryID: string, value: string) => (
+  dispatch,
+  getState: GetState
+) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = fill(query.queryConfig, value)
+
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const removeFuncsAsync = (queryID: string, fields: Field[]) => (
+  dispatch,
+  getState: GetState
+) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = removeFuncs(query.queryConfig, fields)
+
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const applyFuncsToFieldAsync = (
+  queryID: string,
+  fieldFunc: ApplyFuncsToFieldArgs,
+  groupBy?: GroupBy
+) => (dispatch, getState: GetState) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = applyFuncsToField(
+        query.queryConfig,
+        fieldFunc,
+        groupBy
+      )
+
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const chooseTagAsync = (queryID: string, tag: Tag) => (
+  dispatch,
+  getState: GetState
+) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = chooseTag(query.queryConfig, tag)
+
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+interface DBRP {
+  database: string
+  retentionPolicy: string
+}
+
+export const chooseNamespaceAsync = (
+  queryID: string,
+  {database, retentionPolicy}: DBRP
+) => async (dispatch, getState: GetState) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = chooseNamespace(query.queryConfig, {
+        database,
+        retentionPolicy,
+      })
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const chooseMeasurementAsync = (
+  queryID: string,
+  measurement: string
+) => async (dispatch, getState: GetState) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = {
+        ...chooseMeasurement(query.queryConfig, measurement),
+        rawText: getDeep<string>(query, 'queryConfig.rawText', ''),
+      }
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const groupByTagAsync = (queryID: string, tagKey: string) => async (
+  dispatch,
+  getState: GetState
+) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = groupByTag(query.queryConfig, tagKey)
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const toggleTagAcceptanceAsync = (queryID: string) => async (
+  dispatch,
+  getState: GetState
+) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = toggleTagAcceptance(query.queryConfig)
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const addInitialFieldAsync = (
+  queryID: string,
+  field: Field,
+  groupBy: GroupBy
+) => async (dispatch, getState: GetState) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = addInitialField(query.queryConfig, field, groupBy)
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const editQueryStatus = (queryID: string, status: Status) => async (
+  dispatch,
+  getState: GetState
+) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = {...query.queryConfig, status}
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export const timeShiftAsync = (queryID: string, shift: TimeShift) => async (
+  dispatch,
+  getState: GetState
+) => {
+  const {
+    cellEditorOverlay: {queryDrafts},
+  } = getState()
+
+  const updatedQueryDrafts = queryDrafts.map(query => {
+    if (query.id === queryID) {
+      const nextQueryConfig = timeShift(query.queryConfig, shift)
+      return {...query, queryConfig: nextQueryConfig}
+    }
+    return query
+  })
+  dispatch(updateQueryDrafts(updatedQueryDrafts))
+}
+
+export type QueryConfigActions = typeof queryConfigActions & {
+  editRawTextAsync?: (text: string) => Promise<void>
+}
+
+export const queryConfigActions = {
+  fill: fillAsync,
+  timeShift: timeShiftAsync,
+  chooseTag: chooseTagAsync,
+  groupByTag: groupByTagAsync,
+  groupByTime: groupByTimeAsync,
+  toggleField: toggleFieldAsync,
+  removeFuncs: removeFuncsAsync,
+  addInitialField: addInitialFieldAsync,
+  applyFuncsToField: applyFuncsToFieldAsync,
+  toggleTagAcceptance: toggleTagAcceptanceAsync,
+  chooseNamespace: chooseNamespaceAsync,
+  chooseMeasurement: chooseMeasurementAsync,
+}