+++ ui/src/shared/components/TimeMachine/TimeMachineBottom.tsx
@@ -1,0 +1,11 @@
+import React, {ReactNode, SFC} from 'react'
+
+interface Props {
+  children: ReactNode
+}
+
+const TimeMachineBottom: SFC<Props> = ({children}) => (
+  <div className="deceo--bottom">{children}</div>
+)
+
+export default TimeMachineBottom